# docs/swagger.yaml
basePath: /
definitions:
...
  model.KBMetadata:
    additionalProperties: true
    type: object
...
  model.KnowledgeCreateRequest:
    properties:
      category:
        enum:
        - vulnerability
        - security
        - tool
        - guide
        - best_practice
        type: string
      content:
        minLength: 1
        type: string
      metadata:
        $ref: '#/definitions/model.KBMetadata'
      severity:
        enum:
        - low
        - medium
        - high
        - critical
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 255
        minLength: 1
        type: string
      type:
        enum:
        - article
        - tutorial
        - faq
        - solution
        - reference
        type: string
    required:
    - category
    - content
    - title
    - type
    type: object
...
  model.PaginationResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  model.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  model.TagUpdateRequest:
    properties:
      category:
        enum:
        - vulnerability
        - security
        - tool
        - language
        - framework
        type: string
      color:
        type: string
      description:
        type: string
      display_name:
        maxLength: 100
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@vulnark.com
    name: VulnArk Team
    url: https://vulnark.com
  description: VulnArk漏洞管理系统API文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://vulnark.com/terms
  title: VulnArk API
  version: 1.0.0
paths:
...
  /api/v1/knowledge:
    get:
      description: 获取知识库列表（分页）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 分类
        in: query
        name: category
        type: string
      - description: 类型
        in: query
        name: type
        type: string
      - description: 严重程度
        in: query
        name: severity
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 作者ID
        in: query
        name: author_id
        type: string
      - description: 标签
        in: query
        name: tags
        type: string
      - description: 排序字段
        in: query
        name: sort_by
        type: string
      - description: 是否降序
        in: query
        name: sort_desc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取知识库列表
      tags:
      - 知识库管理
    post:
      consumes:
      - application/json
      description: 创建新的知识库条目
      parameters:
      - description: 创建知识库请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.KnowledgeCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.KnowledgeBase'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建知识库
      tags:
      - 知识库管理
  /api/v1/knowledge/{id}:
    delete:
      description: 删除知识库
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除知识库
      tags:
      - 知识库管理
    get:
      description: 根据ID获取知识库详情
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.KnowledgeBase'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取知识库详情
      tags:
      - 知识库管理
    put:
      consumes:
      - application/json
      description: 更新知识库信息
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.KnowledgeUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新知识库
      tags:
      - 知识库管理
  /api/v1/knowledge/{id}/archive:
    post:
      description: 归档知识库
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 归档知识库
      tags:
      - 知识库管理
  /api/v1/knowledge/{id}/like:
    post:
      description: 为知识库点赞
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 点赞知识库
      tags:
      - 知识库管理
  /api/v1/knowledge/{id}/publish:
    post:
      description: 发布知识库
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 发布知识库
      tags:
      - 知识库管理
  /api/v1/knowledge/{id}/related:
    get:
      description: 获取与指定知识库相关的其他知识库
      parameters:
      - description: 知识库ID
        in: path
        name: id
        required: true
        type: integer
      - default: 5
        description: 限制数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.KnowledgeBase'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取相关知识库
      tags:
      - 知识库管理
  /api/v1/knowledge/popular:
    get:
      description: 获取热门知识库列表
      parameters:
      - default: 10
        description: 限制数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.KnowledgeBase'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取热门知识库
      tags:
      - 知识库管理
  /api/v1/knowledge/recent:
    get:
      description: 获取最新知识库列表
      parameters:
      - default: 10
        description: 限制数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.KnowledgeBase'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取最新知识库
      tags:
      - 知识库管理
  /api/v1/knowledge/search:
    get:
      description: 搜索知识库内容
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        required: true
        type: string
      - default: 10
        description: 限制数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.KnowledgeBase'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      summary: 搜索知识库
      tags:
      - 知识库管理
  /api/v1/knowledge/stats:
    get:
      description: 获取知识库统计信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.KnowledgeStats'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取知识库统计
      tags:
      - 知识库管理
...
  /api/v1/tags:
    get:
      description: 获取所有标签列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Tag'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取标签列表
      tags:
      - 标签管理
    post:
      consumes:
      - application/json
      description: 创建新的标签
      parameters:
      - description: 创建标签请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.TagCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Tag'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建标签
      tags:
      - 标签管理
  /api/v1/tags/{id}:
    delete:
      description: 删除标签
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除标签
      tags:
      - 标签管理
    get:
      description: 根据ID获取标签详情
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Tag'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取标签详情
      tags:
      - 标签管理
    put:
      consumes:
      - application/json
      description: 更新标签信息
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.TagUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新标签
      tags:
      - 标签管理
  /api/v1/tags/popular:
    get:
      description: 获取热门标签列表
      parameters:
      - default: 20
        description: 限制数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Tag'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取热门标签
      tags:
      - 标签管理
  /api/v1/templates:
    get:
      description: 获取所有模板列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Template'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取模板列表
      tags:
      - 模板管理
    post:
      consumes:
      - application/json
      description: 创建新的模板
      parameters:
      - description: 创建模板请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.TemplateCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Template'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建模板
      tags:
      - 模板管理
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"